backend/requirements.txt
plaintext
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose